Breakdown scheduler integration tests into a base class and extensions
